Config docker-compose

+++ front-end/Admin/src/sections/products/product-sizes-and-prices.jsx
@@ -1,5 +1,5 @@
 import { useState } from "react";
-import { Box, Button, TextField, Typography } from "@mui/material";
+import { Box, Button, Grid, TextField, Typography } from "@mui/material";
 import productSizeAPI from "src/services/API/productSizeAPI";
 import { CustomAutocomplete } from "src/components/input/CustomAutoComplete";
 import { toast } from "react-toastify";
@@ -8,7 +8,7 @@ import productAPI from "src/services/API/productAPI";
 import { useMutation } from "@tanstack/react-query";
 import LoadingComp from "src/components/loading/LoadingComp";
 
-const SizeAndPrice = ({ productData, details, setDetails }) => {
+const SizeAndPrice = ({ details, setDetails }) => {
   const [newSizeName, setNewSizeName] = useState("");
   const [newSizePrice, setNewSizePrice] = useState("");
   const [newSizeSalePrice, setNewSizeSalePrice] = useState("");
@@ -26,9 +26,34 @@ const SizeAndPrice = ({ productData, details, setDetails }) => {
     },
   });
 
+  const handleAdd = () => {
+    setDetails((prev) => {
+      if (prev.find((sz) => sz.sizeName === newSizeName)) {
+        toast.error("Size already exists");
+        return prev;
+      }
+
+      if (newSizeName === "" || newSizePrice === "") {
+        toast.error("Please fill in all fields");
+        return prev;
+      }
+      return [
+        ...prev,
+        {
+          sizeName: newSizeName,
+          price: newSizePrice,
+          salePrice: newSizeSalePrice,
+        },
+      ];
+    });
+    setNewSizeName("");
+    setNewSizePrice("");
+    setNewSizeSalePrice("");
+  };
+
   const handleConfirmDelete = () => {
     setOpenConfirmDelete(null);
-    if (openConfirmDelete) {
+    if (openConfirmDelete && openConfirmDelete.productDetailsId) {
       deleteDetails(openConfirmDelete, {
         onSuccess: () => {
           setDetails((prev) => {
@@ -40,6 +65,13 @@ const SizeAndPrice = ({ productData, details, setDetails }) => {
           toast.success("Delete size successfully");
         },
       });
+    } else {
+      setDetails((prev) => {
+        return prev.filter(
+          (item) => item.sizeName !== openConfirmDelete.sizeName
+        );
+      });
+      toast.success("Delete size successfully");
     }
   };
 
@@ -50,137 +82,135 @@ const SizeAndPrice = ({ productData, details, setDetails }) => {
           <Typography variant="h6" sx={{ marginBottom: 2 }}>
             Sizes:
           </Typography>
-          {details.map((sz, index) => (
-            <Box
+          {details?.map((sz, index) => (
+            <Grid
+              container
               key={index}
+              gap={2}
               sx={{
-                display: "flex",
-                alignItems: "center",
-                gap: 2,
-                marginBottom: 2,
+                mb: 6,
               }}
             >
-              <CustomAutocomplete
-                current={sz.sizeName}
-                label={"Size"}
-                labelKey={"sizeName"}
-                queryFn={queryFn}
-                onInputChange={(val) =>
-                  setDetails((prev) =>
-                    prev.map((item) =>
-                      item.productDetailsId === sz.productDetailsId
-                        ? { ...item, sizeName: val }
-                        : item
+              <Grid item xs={12} md={5} sx={{ width: "100%" }}>
+                <CustomAutocomplete
+                  current={sz.sizeName}
+                  label={"Size"}
+                  labelKey={"sizeName"}
+                  queryFn={queryFn}
+                  sx={{ width: "100%" }}
+                  onInputChange={(val) =>
+                    setDetails((prev) =>
+                      prev.map((item) =>
+                        item.productDetailsId === sz.productDetailsId
+                          ? { ...item, sizeName: val.value }
+                          : item
+                      )
                     )
-                  )
-                }
-              />
+                  }
+                />
+              </Grid>
 
-              <TextField
-                label="Size Price"
-                value={sz.price || ""}
-                fullWidth
-                onChange={(e) =>
-                  setDetails((prev) => {
-                    return prev.map((item) =>
-                      item.productDetailsId === sz.productDetailsId
-                        ? { ...item, price: Number(e.target.value) }
-                        : item
-                    );
-                  })
-                }
-              />
-              <TextField
-                label="Sale Price"
-                value={sz.salePrice || ""}
-                fullWidth
-                onChange={(e) =>
-                  setDetails((prev) => {
-                    return prev.map((item) =>
-                      item.productDetailsId === sz.productDetailsId
-                        ? { ...item, salePrice: Number(e.target.value) }
-                        : item
-                    );
-                  })
-                }
-              />
-              <Button
-                variant="contained"
-                color="secondary"
-                onClick={() => setOpenConfirmDelete(sz)}
-              >
-                Delete
-              </Button>
-            </Box>
+              <Grid item sm={12} md={5} sx={{ display: "flex", gap: 2 }}>
+                <TextField
+                  label="Size Price"
+                  value={sz.price || ""}
+                  fullWidth
+                  type="number"
+                  onChange={(e) =>
+                    setDetails((prev) => {
+                      return prev.map((item) =>
+                        item.productDetailsId === sz.productDetailsId
+                          ? { ...item, price: Number(e.target.value) }
+                          : item
+                      );
+                    })
+                  }
+                />
+                <TextField
+                  label="Sale Price"
+                  value={sz.salePrice || ""}
+                  type="number"
+                  fullWidth
+                  onChange={(e) =>
+                    setDetails((prev) => {
+                      return prev.map((item) =>
+                        item.productDetailsId === sz.productDetailsId
+                          ? { ...item, salePrice: Number(e.target.value) }
+                          : item
+                      );
+                    })
+                  }
+                />
+              </Grid>
+              <Grid item sm={12} md={1} sx={{ width: "100%" }}>
+                <Button
+                  variant="contained"
+                  color="secondary"
+                  sx={{ width: "100%", height: "100%" }}
+                  onClick={() => setOpenConfirmDelete(sz)}
+                >
+                  Delete
+                </Button>
+              </Grid>
+            </Grid>
           ))}
         </Box>
 
-        <Box sx={{ display: "flex", gap: 2, marginTop: 4 }}>
-          <CustomAutocomplete
-            current={newSizeName}
-            label={"New Size"}
-            labelKey={"sizeName"}
-            queryFn={queryFn}
-            onInputChange={(val) => {
-              console.log("Parent component input change:", val); // Debugging statement
-              setNewSizeName(val);
-            }}
-          />
+        <Grid container gap={2}>
+          <Grid item sm={12} md={5} sx={{ width: "100%" }}>
+            <CustomAutocomplete
+              current={newSizeName}
+              label={"New Size"}
+              sx={{ width: "100%" }}
+              labelKey={"sizeName"}
+              queryFn={queryFn}
+              onInputChange={(val) => {
+                setNewSizeName(val.value);
+              }}
+            />
+          </Grid>
 
-          <TextField
-            label="New Size Price"
-            value={newSizePrice}
-            onChange={(e) => {
-              setNewSizePrice(Number(e.target.value));
-            }}
-            fullWidth
-          />
-          <TextField
-            label="New Size Sale Price"
-            value={newSizeSalePrice}
-            onChange={(e) => {
-              setNewSizeSalePrice(Number(e.target.value));
-            }}
-            fullWidth
-          />
-          <Button
-            variant="contained"
-            color="primary"
-            onClick={() => {
-              setDetails((prev) => {
-                if (prev.find((sz) => sz.sizeName === newSizeName)) {
-                  toast.error("Size already exists");
-                  return prev;
-                }
-                console.log(newSizeName, newSizePrice, newSizeSalePrice);
-
-                if (newSizeName === "" || newSizePrice === "") {
-                  toast.error("Please fill in all fields");
-                  return prev;
-                }
-                return [
-                  ...prev,
-                  {
-                    sizeName: newSizeName,
-                    price: newSizePrice,
-                    salePrice: newSizeSalePrice,
-                  },
-                ];
-              });
-              setNewSizeName("");
-              setNewSizePrice("");
-              setNewSizeSalePrice("");
-            }}
-          >
-            Add
-          </Button>
-        </Box>
+          <Grid item sm={12} md={5} sx={{ display: "flex", gap: 2 }}>
+            <TextField
+              label="New Size Price"
+              value={newSizePrice}
+              type="number"
+              onChange={(e) => {
+                setNewSizePrice(Number(e.target.value));
+              }}
+              fullWidth
+            />
+            <TextField
+              label="New Size Sale Price"
+              value={newSizeSalePrice}
+              type="number"
+              onChange={(e) => {
+                setNewSizeSalePrice(Number(e.target.value));
+              }}
+              fullWidth
+            />
+          </Grid>
+          <Grid item sm={12} md={1} sx={{ width: "100%" }}>
+            <Button
+              variant="contained"
+              color="primary"
+              onClick={handleAdd}
+              sx={{ width: "100%", height: "100%" }}
+            >
+              Add
+            </Button>
+          </Grid>
+        </Grid>
       </Box>
+
+      {/* Confirm delete */}
       <ConfirmDelete
         open={openConfirmDelete ? true : false}
         onClose={() => setOpenConfirmDelete(false)}
         onDelete={handleConfirmDelete}
       />
+
+      {/* Loading component */}
       <LoadingComp isLoading={isDeleting} />
     </Box>
   );
